PROV-1318 protect search index queue processing with a Semaphore

+++ app/models/ca_search_indexing_queue.php
@@ -203,20 +203,28 @@ class ca_search_indexing_queue extends BaseModel {
 	}
 	# ------------------------------------------------------
 	static public function process() {
-		$o_db = new Db();
-		$o_result = $o_db->query("SELECT * FROM ca_search_indexing_queue ORDER BY entry_id");
-		$o_si = new SearchIndexer($o_db);
+		$r_sem = sem_get(ftok(__FILE__, 'SearchIndexingQueue'));
 
-		while($o_result->nextRow()) {
-			$o_si->indexRow(
-				$o_result->get('table_num'), $o_result->get('row_id'),
-				caUnserializeForDatabase($o_result->get('field_data')),
-				(bool)$o_result->get('reindex'), null,
-				caUnserializeForDatabase($o_result->get('changed_fields')),
-				array_merge(caUnserializeForDatabase($o_result->get('options')), array('queueIndexing' => false))
-			);
+		if(sem_acquire($r_sem)) {
+			$o_db = new Db();
+			$o_result = $o_db->query("SELECT * FROM ca_search_indexing_queue ORDER BY entry_id");
+			if($o_result && $o_result->numRows()) {
+				$o_si = new SearchIndexer($o_db);
 
-			$o_db->query('DELETE FROM ca_search_indexing_queue WHERE entry_id=?', $o_result->get('entry_id'));
+				while ($o_result->nextRow()) {
+					$o_si->indexRow(
+						$o_result->get('table_num'), $o_result->get('row_id'),
+						caUnserializeForDatabase($o_result->get('field_data')),
+						(bool)$o_result->get('reindex'), null,
+						caUnserializeForDatabase($o_result->get('changed_fields')),
+						array_merge(caUnserializeForDatabase($o_result->get('options')), array('queueIndexing' => false))
+					);
+
+					$o_db->query('DELETE FROM ca_search_indexing_queue WHERE entry_id=?', $o_result->get('entry_id'));
+				}
+			}
+
+			sem_release($r_sem);
 		}
 	}
 	# ------------------------------------------------------
